in progress

+++ src/test/java/com/emc/mongoose/system/CreateLimitBySizeTest.java
@@ -278,9 +278,9 @@ public class CreateLimitBySizeTest {
 
         testContainerIoTraceLogRecords(stepId, ioTraceRecFunc);
 
-        System.out.println(ioTraceRecCount.toString() + "\n" + expectedCount + "\n" + expectedCount * requiredAccuracy);
+        System.out.println(ioTraceRecCount.toString() + "\n" + expectedCount + "\n" + (float)expectedCount * requiredAccuracy);
 
-        assertEquals(expectedCount, ioTraceRecCount.sum(), expectedCount * requiredAccuracy);
+        assertEquals(expectedCount, ioTraceRecCount.sum(), (float)expectedCount * requiredAccuracy);
 
 //        final List<CSVRecord> items = new ArrayList<>();
 //        try (final BufferedReader br = new BufferedReader(new FileReader(hostItemOutputFile))) {
